+++ packages/moonstone/Scrollable/ScrollButtons.js
@@ -67,6 +67,16 @@ class ScrollButtons extends Component {
 		disabled: PropTypes.bool,
 
 		/**
+		 * When it is `true`, it allows 5 way navigation to the ScrollButtons.
+		 * This value is set by `Scrollable`.
+		 *
+		 * @type {Boolean}
+		 * @default false
+		 * @private
+		 */
+		focusableScrollButtons: PropTypes.bool,
+
+		/**
 		* Sets the hint string read when focusing the next button in the scroll bar.
 		*
 		* @type {String}
@@ -117,6 +127,7 @@ class ScrollButtons extends Component {
 	}
 
 	static defaultProps = {
+		focusableScrollButtons: false,
 		onNextScroll: nop,
 		onPrevScroll: nop
 	}
@@ -249,8 +260,7 @@ class ScrollButtons extends Component {
 	focusOnOppositeScrollButton = (ev, direction) => {
 		const buttonNode = (ev.target === this.nextButtonNodeRef) ? this.prevButtonNodeRef : this.nextButtonNodeRef;
 
-		ev.preventDefault();
-		ev.nativeEvent.stopPropagation();
+		ev.stopPropagation();
 
 		if (!Spotlight.focus(buttonNode)) {
 			Spotlight.move(direction);
@@ -259,16 +269,38 @@ class ScrollButtons extends Component {
 
 	onSpotlight = (ev) => {
 		const
-			{rtl, vertical} = this.props,
-			{keyCode, target} = ev,
-			direction = getDirection(keyCode),
-			fromNextToPrev = (vertical && direction === 'up') || (!vertical && direction === (rtl ? 'right' : 'left')),
-			fromPrevToNext = (vertical && direction === 'down') || (!vertical && direction === (rtl ? 'left' : 'right'));
+			{focusableScrollButtons, rtl, vertical} = this.props,
+			{target} = ev;
 
-		// manually focus the opposite scroll button when 5way pressed
-		if ((fromNextToPrev && target === this.nextButtonNodeRef) ||
-			(fromPrevToNext && target === this.prevButtonNodeRef)) {
-			this.focusOnOppositeScrollButton(ev, direction);
+		// We don't need to navigate manually if `focusableScrollButtons` is `false`
+		if (focusableScrollButtons) {
+			const
+				direction = getDirection(ev.keyCode),
+				fromNextToPrev = (vertical && direction === 'up') || (!vertical && direction === (rtl ? 'right' : 'left')),
+				fromPrevToNext = (vertical && direction === 'down') || (!vertical && direction === (rtl ? 'left' : 'right'));
+
+			// manually focus the opposite scroll button when 5way pressed
+			if ((fromNextToPrev && target === this.nextButtonNodeRef) ||
+				(fromPrevToNext && target === this.prevButtonNodeRef)) {
+				this.focusOnOppositeScrollButton(ev, direction);
+			}
+		} else {
+			// If it is vertical `Scrollable`, move focus to the left for ltr or to the right for rtl
+			// If is is horizontal `Scrollable`, move focus to the up
+			const direction = !vertical && 'up' || rtl && 'right' || 'left';
+
+			if (Spotlight.getPointerMode()) {
+				// When changing from "pointer" mode to "5way key" mode,
+				// a pointer is hidden and a last focused item get focused after 30ms.
+				// To make sure the content in `VirtualList` or `Scroller` to be focused after that, we used 50ms.
+				setTimeout(() => {
+					if (Spotlight.getCurrent() === target) {
+						Spotlight.move(direction);
+					}
+				}, 50);
+			} else if (Spotlight.getCurrent() === target) {
+				Spotlight.move(direction);
+			}
 		}
 	}
 
@@ -279,11 +311,16 @@ class ScrollButtons extends Component {
 
 	onKeyDownPrev = (ev) => {
 		const
+			{focusableScrollButtons} = this.props,
 			{nextButtonDisabled} = this.state,
 			{keyCode} = ev;
 
 		if (isPageDown(keyCode) && !nextButtonDisabled) {
-			Spotlight.focus(this.nextButtonNodeRef);
+			if (focusableScrollButtons) {
+				Spotlight.focus(this.nextButtonNodeRef);
+			} else {
+				this.onClickNext(ev);
+			}
 		} else if (isPageUp(keyCode)) {
 			this.onClickPrev(ev);
 		}
@@ -291,11 +328,16 @@ class ScrollButtons extends Component {
 
 	onKeyDownNext = (ev) => {
 		const
+			{focusableScrollButtons} = this.props,
 			{prevButtonDisabled} = this.state,
 			{keyCode} = ev;
 
 		if (isPageUp(keyCode) && !prevButtonDisabled) {
-			Spotlight.focus(this.prevButtonNodeRef);
+			if (focusableScrollButtons) {
+				Spotlight.focus(this.prevButtonNodeRef);
+			} else {
+				this.onClickPrev(ev);
+			}
 		} else if (isPageDown(keyCode)) {
 			this.onClickNext(ev);
 		}
@@ -340,6 +382,7 @@ class ScrollButtons extends Component {
 				onSpotlightDown={this.onSpotlight}
 				onSpotlightLeft={this.onSpotlight}
 				onSpotlightRight={this.onSpotlight}
+				onSpotlightUp={this.onSpotlight}
 				onUp={this.onUp}
 				ref={this.initPrevButtonRef}
 			>
@@ -356,6 +399,7 @@ class ScrollButtons extends Component {
 				onDown={this.onDownNext}
 				onHoldPulse={this.onHoldPulseNext}
 				onKeyDown={this.onKeyDownNext}
+				onSpotlightDown={this.onSpotlight}
 				onSpotlightLeft={this.onSpotlight}
 				onSpotlightRight={this.onSpotlight}
 				onSpotlightUp={this.onSpotlight}