+++ packages/moonstone/internal/Picker/Picker.js
@@ -1,10 +1,11 @@
-import {forward} from '@enact/core/handle';
+import {forward, stopImmediate} from '@enact/core/handle';
 import clamp from 'ramda/src/clamp';
 import equals from 'ramda/src/equals';
 import {is} from '@enact/core/keymap';
-import {Job} from '@enact/core/util';
+import {cap, Job} from '@enact/core/util';
 import React from 'react';
 import ReactDOM from 'react-dom';
+import platform from '@enact/core/platform';
 import PropTypes from 'prop-types';
 import Touchable from '@enact/ui/Touchable';
 import shouldUpdate from 'recompose/shouldUpdate';
@@ -179,7 +180,7 @@ const PickerBase = class extends React.Component {
 
 		/**
 		 * When `true`, the Picker is shown as disabled and does not generate `onChange`
-		 * [events]{@glossary event}.
+		 * [events]{@link /docs/developer-guide/glossary/#event}.
 		 *
 		 * @type {Boolean}
 		 * @public
@@ -397,6 +398,9 @@ const PickerBase = class extends React.Component {
 		if (this.props.joined) {
 			this.containerRef.addEventListener('wheel', this.handleWheel);
 		}
+		if (platform.webos) {
+			this.containerRef.addEventListener('webOSVoice', this.handleVoice);
+		}
 	}
 
 	componentWillReceiveProps (nextProps) {
@@ -427,6 +431,9 @@ const PickerBase = class extends React.Component {
 		this.throttleWheelDec.stop();
 		if (this.props.joined) {
 			this.containerRef.removeEventListener('wheel', this.handleWheel);
+		}
+		if (platform.webos) {
+			this.containerRef.removeEventListener('webOSVoice', this.handleVoice);
 		}
 	}
 
@@ -627,45 +634,66 @@ const PickerBase = class extends React.Component {
 
 	handleDecKeyDown = (ev) => {
 		const {keyCode} = ev;
-		const {
-			onPickerSpotlightDown,
-			onPickerSpotlightLeft,
-			onPickerSpotlightRight,
-			onPickerSpotlightUp,
-			orientation,
-			step
-		} = this.props;
+		const direction = getDirection(keyCode);
 
-		if (isDown(keyCode) && onPickerSpotlightDown) {
-			onPickerSpotlightDown(ev);
-		} else if (isLeft(keyCode) && onPickerSpotlightLeft) {
-			onPickerSpotlightLeft(ev);
-		} else if (isRight(keyCode) && onPickerSpotlightRight && (orientation === 'vertical' || this.hasReachedBound(step))) {
-			onPickerSpotlightRight(ev);
-		} else if (isUp(keyCode) && onPickerSpotlightUp && (orientation === 'horizontal' || this.hasReachedBound(step))) {
-			onPickerSpotlightUp(ev);
+		if (direction) {
+			const {orientation, step} = this.props;
+
+			if (
+				!this.hasReachedBound(step) &&
+				(
+					isRight(keyCode) && orientation === 'horizontal' ||
+					isUp(keyCode) && orientation === 'vertical'
+				)
+			) {
+				ev.preventDefault();
+				// prevent default spotlight behavior
+				stopImmediate(ev);
+				// set the pointer mode to false on keydown
+				Spotlight.setPointerMode(false);
+				Spotlight.focus(this.containerRef.querySelector(`.${css.incrementer}`));
+			} else {
+				forward(`onPickerSpotlight${cap(direction)}`, ev, this.props);
+			}
 		}
 	}
 
 	handleIncKeyDown = (ev) => {
 		const {keyCode} = ev;
-		const {
-			onPickerSpotlightDown,
-			onPickerSpotlightLeft,
-			onPickerSpotlightRight,
-			onPickerSpotlightUp,
-			orientation,
-			step
-		} = this.props;
+		const direction = getDirection(keyCode);
 
-		if (isDown(keyCode) && onPickerSpotlightDown && (orientation === 'horizontal' || this.hasReachedBound(step * -1))) {
-			onPickerSpotlightDown(ev);
-		} else if (isLeft(keyCode) && onPickerSpotlightLeft && (orientation === 'vertical' || this.hasReachedBound(step * -1))) {
-			onPickerSpotlightLeft(ev);
-		} else if (isRight(keyCode) && onPickerSpotlightRight) {
-			onPickerSpotlightRight(ev);
-		} else if (isUp(keyCode) && onPickerSpotlightUp) {
-			onPickerSpotlightUp(ev);
+		if (direction) {
+			const {orientation, step} = this.props;
+
+			if (
+				!this.hasReachedBound(step * -1) &&
+				(
+					isLeft(keyCode) && orientation === 'horizontal' ||
+					isDown(keyCode) && orientation === 'vertical'
+				)
+			) {
+				ev.preventDefault();
+				// prevent default spotlight behavior
+				stopImmediate(ev);
+				// set the pointer mode to false on keydown
+				Spotlight.setPointerMode(false);
+				Spotlight.focus(this.containerRef.querySelector(`.${css.decrementer}`));
+			} else {
+				forward(`onPickerSpotlight${cap(direction)}`, ev, this.props);
+			}
+		}
+	}
+
+	handleVoice = (ev) => {
+		const voiceIndex = ev && ev.detail && typeof ev.detail.matchedIndex !== 'undefined' && Number(ev.detail.matchedIndex);
+
+		if (Number.isInteger(voiceIndex)) {
+			const {max, min, onChange, value} = this.props;
+			const voiceValue = min + voiceIndex;
+			if (onChange && voiceValue >= min && voiceValue <= max && voiceValue !== value) {
+				onChange({value: voiceValue});
+				ev.preventDefault();
+			}
 		}
 	}
 
@@ -704,24 +732,22 @@ const PickerBase = class extends React.Component {
 	}
 
 	calcButtonLabel (next, valueText) {
-		if (!this.props.joined) {
-			const {decrementAriaLabel, incrementAriaLabel} = this.props;
-			let label = next ? incrementAriaLabel : decrementAriaLabel;
+		const {decrementAriaLabel, incrementAriaLabel} = this.props;
+		let label = next ? incrementAriaLabel : decrementAriaLabel;
 
-			if (label != null) {
-				return label;
-			}
-
-			return `${valueText} ${next ? $L('next item') : $L('previous item')}`;
+		if (label != null) {
+			return label;
 		}
+
+		return `${valueText} ${next ? $L('next item') : $L('previous item')}`;
 	}
 
 	calcDecrementLabel (valueText) {
-		return this.calcButtonLabel(this.props.reverse, valueText);
+		return !this.props.joined ? this.calcButtonLabel(this.props.reverse, valueText) : null;
 	}
 
 	calcIncrementLabel (valueText) {
-		return this.calcButtonLabel(!this.props.reverse, valueText);
+		return !this.props.joined ? this.calcButtonLabel(!this.props.reverse, valueText) : null;
 	}
 
 	calcAriaLabel (valueText) {
@@ -749,15 +775,16 @@ const PickerBase = class extends React.Component {
 		const {active} = this.state;
 		const {
 			'aria-valuetext': ariaValueText,
-			'data-webos-voice-group-label': voiceGroupLabel,
 			noAnimation,
 			children,
+			'data-webos-voice-group-label': voiceGroupLabel,
 			disabled,
 			id,
 			index,
 			joined,
 			onSpotlightDisappear,
 			orientation,
+			reverse,
 			spotlightDisabled,
 			step,
 			width,
@@ -777,7 +804,6 @@ const PickerBase = class extends React.Component {
 		delete rest.onPickerSpotlightLeft;
 		delete rest.onPickerSpotlightRight;
 		delete rest.onPickerSpotlightUp;
-		delete rest.reverse;
 		delete rest.value;
 		delete rest.wrap;
 
@@ -811,6 +837,8 @@ const PickerBase = class extends React.Component {
 				aria-disabled={disabled}
 				aria-label={this.calcAriaLabel(valueText)}
 				className={classes}
+				data-webos-voice-group-label={voiceGroupLabel}
+				data-webos-voice-intent="Select"
 				disabled={disabled}
 				onBlur={this.handleBlur}
 				onFocus={this.handleFocus}
@@ -825,6 +853,7 @@ const PickerBase = class extends React.Component {
 					aria-label={this.calcIncrementLabel(valueText)}
 					className={css.incrementer}
 					data-webos-voice-group-label={voiceGroupLabel}
+					data-webos-voice-label={joined ? this.calcButtonLabel(!reverse, valueText) : null}
 					disabled={incrementerDisabled}
 					hidden={reachedEnd}
 					icon={incrementIcon}
@@ -860,6 +889,7 @@ const PickerBase = class extends React.Component {
 					aria-label={this.calcDecrementLabel(valueText)}
 					className={css.decrementer}
 					data-webos-voice-group-label={voiceGroupLabel}
+					data-webos-voice-label={joined ? this.calcButtonLabel(reverse, valueText) : null}
 					disabled={decrementerDisabled}
 					hidden={reachedStart}
 					icon={decrementIcon}